syntax etc. fixes for sparc
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@2931 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/cbfa.c
@@ -234,9 +234,9 @@ void PrintCBFA()
 	  limit = (long)cbfa->entries + CBFABlockSize;
 	}
 	numEntry++;
-	fprintf(output, "Entry no %d: 0x%x\n", numEntry, current);
+	fprintf(output, "Entry no %d: 0x%x\n", (int)numEntry, (int)current);
 	fprintf(output, "  theStruct: 0x%08x: 0x%08x ", 
-		&current->theStruct, current->theStruct);
+		(int)&current->theStruct, (int)current->theStruct);
 	if ((current->theStruct) 
 	    && inBetaHeap((ref(Object))(current->theStruct))
 	    && inBetaHeap(current->theStruct->iOrigin)){
